Bug 1380540: support extraLines property
On Linux, Firefox renders scrollbars all the time, rather than fading
them in and out on hover as occurs on other platforms.  Thus if there is
horizontal scrolling, the last line of a log is always obscured by a
scroll bar.  Typically that last line is the most important one!

This allows users to work around this issue by including extra, blank
lines at the end of the log output, even while streaming.